ASCII strings support added.
there was a problem when application were asking non UTF-8 string from clipboard
now it can ask ASCII-strings as well as UTF-8 strings.

TODO: now we work with UTF-8 strings like if it always would consist of 8-bits
characters, but it wouldn't always be so, because there would be
16-/32-bits characters in it. So we probably need to work with this
strings properly.